backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.6.1
python-multipart==0.0.17
slowapi==0.1.9
pgvector==0.3.5
python-dotenv==1.0.1
onnxruntime==1.17.3
# Phase 3 — Face Recognition
insightface
opencv-python-headless
numpy==1.26.4
